+++ src/test/java/com/hivemq/bootstrap/netty/ChannelDependenciesTest.java
@@ -37,8 +37,6 @@ import com.hivemq.mqtt.handler.publish.DropOutgoingPublishesHandler;
 import com.hivemq.mqtt.handler.publish.MessageExpiryHandler;
 import com.hivemq.mqtt.handler.publish.PublishUserEventReceivedHandler;
 import com.hivemq.mqtt.handler.publish.ReturnMessageIdToPoolHandler;
-import com.hivemq.mqtt.handler.publish.qos.QoSReceiverHandler;
-import com.hivemq.mqtt.handler.publish.qos.QoSSenderHandler;
 import com.hivemq.mqtt.handler.subscribe.SubscribeHandler;
 import com.hivemq.mqtt.handler.unsubscribe.UnsubscribeHandler;
 import com.hivemq.security.ssl.SslParameterHandler;
@@ -77,12 +75,6 @@ public class ChannelDependenciesTest {
 
     @Mock
     private UnsubscribeHandler unsubscribeHandler;
-
-    @Mock
-    private QoSReceiverHandler qoSReceiverHandler;
-
-    @Mock
-    private QoSSenderHandler qoSSenderHandler;
 
     @Mock
     private ChannelGroup channelGroup;
@@ -172,8 +164,6 @@ public class ChannelDependenciesTest {
                 () -> subscribeHandler,
                 () -> publishUserEventReceivedHandler,
                 () -> unsubscribeHandler,
-                () -> qoSReceiverHandler,
-                () -> qoSSenderHandler,
                 channelGroup,
                 fullConfigurationService,
                 globalTrafficShapingHandler,
@@ -210,8 +200,6 @@ public class ChannelDependenciesTest {
         assertNotNull(channelDependencies.getSubscribeHandler());
         assertNotNull(channelDependencies.getPublishUserEventReceivedHandler());
         assertNotNull(channelDependencies.getUnsubscribeHandler());
-        assertNotNull(channelDependencies.getQoSSenderHandler());
-        assertNotNull(channelDependencies.getQoSReceiverHandler());
         assertNotNull(channelDependencies.getChannelGroup());
         assertNotNull(channelDependencies.getConfigurationService());
         assertNotNull(channelDependencies.getGlobalTrafficShapingHandler());
